+++ services/payroll_engine/src/main/java/com/fenomina/payroll_engine/service/liquidacion/LiquidacionInteresesServiceImpl.java
@@ -132,7 +132,7 @@ public class LiquidacionInteresesServiceImpl implements LiquidacionInteresesServ
                             .filter(d -> d.getFkConcepNominaId() != null &&
                                     d.getFkConcepNominaId().equals(1L))
                             .mapToInt(d -> d.getCantidadConcept() != null
-                                    ? d.getCantidadConcept() : 0)
+                                    ? d.getCantidadConcept().intValue() : 0)
                             .sum())
                     .sum();
             diasLiquidados = Math.min(diasLiquidados, DIAS_ANIO);
